+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## [v1.0.0-RC1]
 
 - Add [#51]: metadata for output to make it more future proof.
+- Add [#35]: GET method for /matches in REST API.
 
 ## [v0.9.9] - 2022-05-05 Thu